clean: Rename require utility
