+++ routes/diseases.js
@@ -1,9 +1,9 @@
-module.exports = (router) => {
-    const diseasesController = global.diContainer.get('controller.disease');
+var express = require('express');
+var router = express.Router();
+const diseasesController = global.diContainer.get('controller.disease');
 
-    router.get('/', diseasesController.list);
-    router.get('/search', diseasesController.search);
-    router.get('/:id', diseasesController.getById);
+router.get('/', diseasesController.list);
+router.get('/search', diseasesController.search);
+router.get('/:id', diseasesController.getById);
 
-    return router
-};
+module.exports = router;
