+++ routes/init.js
@@ -1,10 +1,9 @@
-var express = require('express');
 const swaggerUi = require('swagger-ui-express'), swaggerDocument = require('../swagger.json');
 const authmid = require('../auth/middleware')
 const baseurl = '/api';
 const v1 = baseurl + '/v1';
 
-module.exports = (app, router) => {
+module.exports = (app) => {
     app.use('/', require('../routes/index'));
     app.use(baseurl + '/', require('../routes/index'));
     app.use(baseurl + '/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerDocument));
@@ -20,6 +19,6 @@ module.exports = (app, router) => {
     app.use(v1 + '/bmi', require('../routes/bmi'));
     app.use(v1 + '/appdb', require('../routes/appdb'));
 
-    app.use(v1 + '/diseases', authmid, require('../routes/diseases')(express.Router()));
+    app.use(v1 + '/diseases', authmid, require('../routes/diseases'));
 }
